+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ HEAD
 
 **To be included in next release**
 
+  - fix an issue in profiles which contained only persistent properties
 1.17.2 (2019-08-16)
 -------------------